Refactor a lot, move into distinct modules
